refactor: centralize security helpers
